refactor: repo `Package`
