+++ apps/io-session-manager-internal/src/controllers/__tests__/info.test.ts
@@ -4,15 +4,15 @@ import * as H from "@pagopa/handler-kit";
 import { makeInfoHandler } from "../info";
 import { httpHandlerInputMocks } from "../__mocks__/handler.mock";
 import { mockInfoService } from "../../services/__mocks__/info.mock";
-import { aPackageInfo } from "../../__mocks__/package.mock";
-import { PackageUtils } from "../../utils/package";
+import { aPackageInfo } from "../../repositories/__mocks__/package.mock";
+import { Package } from "../../repositories/package";
 
 describe("Info handler", () => {
   it("should succeed if the application is healthy", async () => {
     const result = await makeInfoHandler({
       ...httpHandlerInputMocks,
       InfoService: mockInfoService,
-      PackageUtils: {} as PackageUtils,
+      Package: {} as Package,
     })();
 
     expect(result).toMatchObject(E.right(H.successJson(aPackageInfo)));
@@ -26,7 +26,7 @@ describe("Info handler", () => {
     const result = await makeInfoHandler({
       ...httpHandlerInputMocks,
       InfoService: mockInfoService,
-      PackageUtils: {} as PackageUtils,
+      Package: {} as Package,
     })();
 
     expect(result).toMatchObject(
