+++ apps/io-session-manager-internal/src/repositories/__mocks__/package.mock.ts
@@ -1,5 +1,5 @@
 import { vi } from "vitest";
-import { PackageUtils } from "../utils/package";
+import { Package } from "../package";
 
 export const aPackageName = "io-session-manager-internal";
 export const aPackageVersion = "1.0.0";
@@ -8,7 +8,7 @@ export const aPackageInfo = {
   version: aPackageVersion,
 };
 
-export const mockPackageUtils: PackageUtils = {
+export const mockPackage: Package = {
   getValueFromPackageJson: vi.fn().mockReturnValue(aPackageName),
   getCurrentBackendVersion: vi.fn().mockReturnValue(aPackageVersion),
 };
