+++ apps/io-session-manager-internal/src/services/__mocks__/info.mock.ts
@@ -1,7 +1,7 @@
 import { vi } from "vitest";
 import * as E from "fp-ts/Either";
 import { InfoService } from "../info";
-import { aPackageInfo } from "../../__mocks__/package.mock";
+import { aPackageInfo } from "../../repositories/__mocks__/package.mock";
 
 export const mockInfoService: InfoService = {
   getPackageInfo: vi.fn(() => () => Promise.resolve(E.right(aPackageInfo))),
